use features to gate deps properly
