+++ include-idl-cli/Cargo.toml
@@ -5,6 +5,5 @@ edition = "2021"
 
 [dependencies]
 clap = { version = "4.5.7", features = ["derive"] }
-flate2 = "1.0"
-goblin = "0.8.2"
-serde_json = "1.0.117"
+goblin = { version = "0.8.2" }
+include-idl = { path = "../include-idl", features = ["parse"] }
