+++ include-idl/Cargo.toml
@@ -5,5 +5,11 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+shrink = ["flate2"]
+parse = ["flate2", "goblin", "serde_json"]
+
 [dependencies]
-flate2 = "1.0"
+flate2 = { version = "1.0", optional = true }
+goblin = { version = "0.8.2", optional = true }
+serde_json = { version = "1.0.117", optional = true }
